refactor: add `tiger-bin-shared`
