+++ Cargo.toml
@@ -1,6 +1,13 @@
 [workspace]
 
-members = ["ck3-tiger", "vic3-tiger", "imperator-tiger", "utils", "."]
+members = [
+    "ck3-tiger",
+    "vic3-tiger",
+    "imperator-tiger",
+    "tiger-bin-shared",
+    "utils",
+    ".",
+]
 
 [package]
 name = "tiger-lib"
@@ -40,7 +47,6 @@ once_cell = "1.18.0"
 phf = { version = "0.11", features = ["macros"] }
 png = "0.17.6"
 rayon = "1.7.0"
-regex = "1.10"
 serde = { version = "1.0.184", features = ["derive"] }
 serde_json = "1.0.104"
 strum = { version = "0.26", features = ["phf"] }
@@ -52,19 +58,6 @@ ahash = { version = "0.8", features = [
     "std",
     "compile-time-rng",
 ], default-features = false }
-
-[target.'cfg(target_os = "windows")'.dependencies]
-self_update = { version = "0.39", features = [
-    "archive-zip",
-    "compression-zip-deflate",
-] }
-winreg = "0.52"
-
-[target.'cfg(target_os = "linux")'.dependencies]
-self_update = { version = "0.39", features = [
-    "archive-tar",
-    "compression-flate2",
-] }
 
 [profile.bench]
 debug = true
