+++ ck3-tiger/Cargo.toml
@@ -15,16 +15,14 @@ rust-version = "1.70"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-tiger-lib = { version = "0.9.5", path = "..", features = ["ck3"] }
+tiger-lib = { path = "..", features = ["ck3"] }
+tiger-bin-shared = { path = "../tiger-bin-shared", features = ["ck3"] }
 
 anyhow = "1"
 clap = { version = "~4.4", features = ["derive"] }
 console = "0.15.7"
 serde_json = "1.0.104"
 strum = "0.26.0"
-
-[target.'cfg(windows)'.dependencies]
-ansiterm = "0.12.2"
 
 [dev-dependencies]
 serde = { version = "1.0.184", features = ["derive"] }
