+++ src/lib.rs
@@ -65,7 +65,6 @@ pub use crate::report::{
     Severity,
 };
 pub use crate::token::{Loc, Token};
-pub use crate::update::{update, UpdateError};
 
 #[cfg(feature = "ck3")]
 mod ck3;
@@ -110,7 +109,6 @@ mod script_value;
 mod token;
 mod tooltipped;
 mod trigger;
-mod update;
 mod util;
 mod validate;
 mod validator;
