+++ vic3-tiger/Cargo.toml
@@ -15,10 +15,10 @@ rust-version = "1.70"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-tiger-lib = { version = "0.9.5", path = "..", default-features = false, features = ["vic3"] }
+tiger-lib = { path = "..", default-features = false, features = ["vic3"] }
+tiger-bin-shared = { path = "../tiger-bin-shared", default-features = false, features = ["vic3"] }
 
 anyhow = "1"
-clap = { version = "~4.4", features = ["derive"] }
 console = "0.15.7"
 
 [target.'cfg(windows)'.dependencies]
